+++ api/app/c_haines/severity_index.py
@@ -20,8 +20,7 @@ from wps_shared.utils.s3 import object_exists, object_exists_v2
 import wps_shared.utils.time as time_utils
 from wps_shared.weather_models import ModelEnum, ProjectionEnum
 from wps_shared.geospatial.geospatial import WGS84
-from app.jobs.env_canada import get_model_run_hours, adjust_model_day, download, UnhandledPredictionModelType
-from app.jobs.env_canada_utils import get_file_date_part
+
 from wps_shared.utils.s3 import get_client
 from app.c_haines import get_severity_string
 from app.c_haines.c_haines_index import CHainesGenerator
@@ -29,7 +28,8 @@ from app.c_haines import GDALData
 from app.c_haines.object_store import ObjectTypeEnum, generate_full_object_store_path
 from app.c_haines.kml import save_as_kml_to_s3
 from wps_shared import config
-from app.weather_models.process_grib import get_dataset_transform
+from wps_shared.geospatial.geospatial import get_dataset_transform
+from wps_shared.weather_models import adjust_model_day, download, get_env_canada_model_run_hours, get_file_date_part, UnhandledPredictionModelType
 
 
 logger = logging.getLogger(__name__)
@@ -377,7 +377,7 @@ class CHainesSeverityGenerator:
     async def _get_payloads(self, temporary_path) -> Generator[EnvCanadaPayload, None, None]:
         """Iterator that yields the next to process."""
         utc_now = time_utils.get_utc_now()
-        for model_hour in get_model_run_hours(self.model):
+        for model_hour in get_env_canada_model_run_hours(self.model):
             for prediction_hour in model_prediction_hour_iterator(self.model):
                 urls, model_run_timestamp, prediction_timestamp = make_model_run_download_urls(self.model, utc_now, model_hour, prediction_hour)
