+++ api/app/routers/morecast_v2.py
@@ -6,7 +6,7 @@ from app.morecast_v2.forecasts import format_as_wf1_post_forecasts
 from wps_shared.utils.time import vancouver_tz
 from typing import List
 from datetime import date, datetime, time, timedelta, timezone
-from fastapi import APIRouter, Response, Depends, status, HTTPException
+from fastapi import APIRouter, Response, Depends, status
 from fastapi.responses import ORJSONResponse
 from app.auth import auth_with_forecaster_role_required, audit, authentication_required
 from wps_shared.db.crud.grass_curing import get_percent_grass_curing_by_station_for_date_range
@@ -27,7 +27,7 @@ from wps_shared.schemas.morecast_v2 import (
 from wps_shared.schemas.shared import StationsRequest
 from wps_shared.wildfire_one.schema_parsers import transform_morecastforecastoutput_to_weatherindeterminate
 from wps_shared.utils.time import get_hour_20_from_date, get_utc_now
-from app.weather_models.fetch.predictions import fetch_latest_model_run_predictions_by_station_code_and_date_range
+from wps_shared.weather_models.fetch.predictions import fetch_latest_model_run_predictions_by_station_code_and_date_range
 from wps_shared.wildfire_one.wfwx_api import get_auth_header, get_dailies_for_stations_and_date, get_daily_determinates_for_stations_and_date, get_wfwx_stations_from_station_codes
 from app.wildfire_one.wfwx_post_api import WF1_HTTP_ERROR, post_forecasts
 from wps_shared.utils.redis import clear_cache_matching
