+++ api/app/tests/weather_models/test_grib_processing.py
@@ -8,7 +8,7 @@ from affine import Affine
 from pyproj import CRS
 from osgeo import gdal
 from wps_shared.geospatial.geospatial import NAD83_CRS
-import app.weather_models.process_grib as process_grib
+import wps_jobs.wps_jobs.weather_model_jobs.utils.process_grib as process_grib
 
 logger = logging.getLogger(__name__)
