+++ api/app/tests/weather_models/test_noaa_gfs.py
@@ -3,15 +3,15 @@
 import os
 import logging
 from datetime import datetime, timezone
-from app.jobs import common_model_fetchers
 from app.tests.weather_models.test_models_common import MockResponse, shape, mock_get_model_run_predictions
-import wps_shared.utils.time as time_utils
 import pytest
 import requests
 from geoalchemy2.shape import from_shape
-from wps_shared.db.models.weather_models import PredictionModel, PredictionModelGridSubset, PredictionModelRunTimestamp, ProcessedModelRunUrl
-from app.jobs import noaa
+from wps_jobs.wps_jobs.weather_model_jobs import common_model_fetchers
 import wps_shared.db.crud.weather_models
+from wps_shared.db.models.weather_models import PredictionModel, PredictionModelGridSubset, PredictionModelRunTimestamp, ProcessedModelRunUrl
+from wps_jobs.wps_jobs.weather_model_jobs import noaa
+import wps_shared.utils.time as time_utils
 
 
 logger = logging.getLogger(__name__)
