+++ wps_jobs/pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "wps-shared"
+name = "wps-jobs"
 version = "0.1.0"
 description = ""
 authors = ["Darren Boss <darren.boss@gov.bc.ca>"]
@@ -7,22 +7,17 @@ readme = "README.md"
 
 [tool.poetry.dependencies]
 python = ">=3.12.3"
-sqlalchemy = "^2"
-aiohttp = "^3"
-requests = "^2"
-python-decouple = "^3"
-geopandas = "^1.0.1"
-shapely = "^2.0.5"
-pydantic = "^2.10.6"
-aiobotocore = "^2.0.0"
-geoalchemy2 = "^0"
+wps-shared = {path = "../wps_shared"}
+scipy = "^1.15.2"
+scikit-learn = "^1.6.1"
+numpy = "^2.2.3"
+sqlalchemy = "^2.0.38"
+redis = "^5.2.1"
+psycopg2-binary = "^2.9.10"
+asyncpg = "^0.30.0"
 affine = "^2.4.0"
 
 
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
-
-[tool.ruff]
-line-length = 185
-lint.ignore = ["E712", "F401"]
