+++ wps_jobs/wps_jobs/__init__.py
@@ -1,0 +1,1 @@
+""" Module for wps jobs"""
