+++ wps_jobs/wps_jobs/weather_model_jobs/common_model_fetchers.py
@@ -1,7 +1,5 @@
-import os
 from typing import List
 import logging
-import requests
 import numpy
 from datetime import datetime, timedelta, timezone
 from pyproj import Geod
@@ -16,14 +14,12 @@ from wps_shared.db.crud.weather_models import (
     delete_weather_station_model_predictions,
     delete_model_run_predictions,
 )
-from app.weather_models.machine_learning import StationMachineLearning
+from wps_jobs.weather_models.interpolate import construct_interpolated_noon_prediction, interpolate_between_two_points
+from wps_jobs.weather_model_jobs.machine_learning import StationMachineLearning
 from wps_shared.weather_models import ModelEnum
-from app.weather_models.interpolate import construct_interpolated_noon_prediction, interpolate_between_two_points
 from wps_shared.schemas.stations import WeatherStation
 from wps_shared.logging import configure_logging
 import wps_shared.utils.time as time_utils
-from wps_shared.utils.redis import create_redis
-from wps_shared import config
 from wps_shared.stations import get_stations_synchronously
 from wps_shared.db.models.weather_models import ProcessedModelRunUrl, PredictionModelRunTimestamp, WeatherStationModelPrediction, ModelRunPrediction
 import wps_shared.db.database
@@ -38,84 +34,8 @@ logger = logging.getLogger(__name__)
 # Keys for weather variables that require interpolation between 1800 and 2100
 SCALAR_MODEL_VALUE_KEYS_FOR_INTERPOLATION = ("tmp_tgl_2", "rh_tgl_2", "wind_tgl_10", "apcp_sfc_0")
 
-
-class UnhandledPredictionModelType(Exception):
-    """Exception raised when an unknown model type is encountered."""
-
-
 class CompletedWithSomeExceptions(Exception):
     """Exception raised when processing completed, but there were some non critical exceptions"""
-
-
-def download(url: str, path: str, config_cache_var: str, model_name: str, config_cache_expiry_var=None) -> str:
-    """
-    Download a file from a url.
-    NOTE: was using wget library initially, but has the drawback of not being able to control where the
-    temporary files are stored. This is problematic, as giving the application write access to /app
-    is a security concern.
-    TODO: Would be nice to make this an async
-    """
-    if model_name == "GFS":
-        original_filename = os.path.split(url)[-1]
-        # NOTE: This is a very not-ideal way to interpolate the filename.
-        # The original_filename that we get from the url is too long and must be condensed.
-        # It also has multiple '.' chars in the URL that must be removed for the filename to be valid.
-        # As long as NOAA's API remains unchanged, we'll have all the info we need (run datetimes,
-        # projections, etc.) in the first 81 characters of original_filename.
-        # An alternative would be to build out a regex to look for
-        filename = original_filename[:81].replace(".", "")
-    else:
-        # Infer filename from url.
-        filename = os.path.split(url)[-1]
-    # Construct target location for downloaded file.
-    target = os.path.join(os.getcwd(), path, filename)
-    # Get the file.
-    # We don't strictly need to use redis - but it helps a lot when debugging on a local machine, it
-    # saves having to re-download the file all the time.
-    # It also save a lot of bandwidth in our dev environment, where we have multiple workers downloading
-    # the same files over and over.
-    if config.get(config_cache_var) == "True":
-        cache = create_redis()
-        try:
-            cached_object = cache.get(url)
-        except Exception as error:
-            cached_object = None
-            logger.error(error)
-    else:
-        cached_object = None
-        cache = None
-    if cached_object:
-        logger.info("Cache hit %s", url)
-        # Store the cached object in a file
-        with open(target, "wb") as file_object:
-            # Write the file.
-            file_object.write(cached_object)
-    else:
-        logger.info("Downloading %s", url)
-        # It's important to have a timeout on the get, otherwise the call may get stuck for an indefinite
-        # amount of time - there is no default value for timeout. During testing, it was observed that
-        # downloads usually complete in less than a second.
-        response = requests.get(url, timeout=60)
-        # If the response is 200/OK.
-        if response.status_code == 200:
-            # Store the response.
-            with open(target, "wb") as file_object:
-                # Write the file.
-                file_object.write(response.content)
-            # Cache the response
-            if cache:
-                with open(target, "rb") as file_object:
-                    # Cache for 6 hours (21600 seconds)
-                    cache.set(url, file_object.read(), ex=config.get(config_cache_expiry_var, 21600))
-        elif response.status_code == 404:
-            # We expect this to happen frequently - just log for info.
-            logger.info("404 error for %s", url)
-            target = None
-        else:
-            # Raise an exception
-            response.raise_for_status()
-        # Return file location.
-    return target
 
 
 def get_closest_index(coordinate: List, points: List):
