+++ wps_jobs/wps_jobs/weather_model_jobs/utils/linear_model.py
@@ -4,7 +4,7 @@ from typing import List
 from sklearn.exceptions import NotFittedError
 from sklearn.linear_model import LinearRegression
 from collections import defaultdict
-from app.weather_models.sample import Samples
+from wps_jobs.weather_models.sample import Samples
 
 logger = logging.getLogger(__name__)
