+++ wps_jobs/wps_jobs/weather_model_jobs/utils/weather_models.py
@@ -3,13 +3,13 @@ import math
 from typing import List
 from wps_shared.db.models.observations import HourlyActual
 from wps_shared.db.models.weather_models import ModelRunPrediction
-from app.weather_models.interpolate import construct_interpolated_noon_prediction
-from app.weather_models.linear_model import LinearModel
-from app.weather_models.precip_model import PrecipModel
-from app.weather_models.regression_model import RegressionModelProto, model_2_actual_keys
-from app.weather_models.sample import Samples
-from app.weather_models.utils import construct_dictionary_from_list_by_property
-from app.weather_models.wind_direction_model import WindDirectionModel
+from wps_jobs.weather_models.interpolate import construct_interpolated_noon_prediction
+from wps_jobs.weather_models.linear_model import LinearModel
+from wps_jobs.weather_models.precip_model import PrecipModel
+from wps_jobs.weather_models.regression_model import RegressionModelProto, model_2_actual_keys
+from wps_jobs.weather_models.sample import Samples
+from wps_jobs.weather_models.utils import construct_dictionary_from_list_by_property
+from wps_jobs.weather_models.wind_direction_model import WindDirectionModel
 
 logger = logging.getLogger(__name__)
